refactor: 퀘스트 메인 모달 디렉터리 분리
